renamed development-mode.ts
